requirements.txt
# Core dependencies for Speech-to-Text Dictation Tool

# Audio capture library (Python bindings for PortAudio)
PyAudio==0.2.14

# Keyboard listener and controller for hotkeys and text injection
pynput==1.7.6

# Optimized Whisper implementation with GPU support and int8 quantization
# This is the key library that enables running quality models in 2GB VRAM
faster-whisper==1.0.3

# Configuration file handling
PyYAML>=6.0

# System tray icon support
pystray==0.19.5

# Image processing for system tray icon
Pillow>=10.0.0

# Numerical operations for audio processing
numpy>=1.24.0

# Screen automation for element detection and clicking
pyautogui>=0.9.54

# Windows UI Automation (Windows-only, for Phase 3 advanced element detection)
pywinauto>=0.6.8; platform_system=="Windows"

# File system monitoring for auto-reload during development
watchdog>=3.0.0

# Optional: For CUDA GPU acceleration (if you have CUDA installed)
# If you don't have CUDA, faster-whisper will automatically fall back to CPU
# Uncomment the following if you need to install CUDA toolkit dependencies:
# ctranslate2==4.3.1

# Development dependencies
# Install with: pip install -r requirements.txt
# Or install just core deps: pip install PyAudio pynput faster-whisper PyYAML pystray Pillow numpy

# Code formatting
black==25.11.0

# Linting
pylint==4.0.2
flake8==7.3.0

# Type checking
mypy==1.18.2

# Testing
pytest==9.0.0
pytest-cov==7.0.0
